Integrate `ContainerLookupService`, enhance logging, and update configurations
- Add `ContainerLookupService` to `BlockChainService` for dynamic container resolution.
- Update `processActiveSubscriptions` scheduling interval to 1000ms.
- Enhance logging in `ComputationService` and related classes with additional details on container and input data.
- Extract `decodeInputDataToString` to `CommonUtil` for reuse across services.
- Update Docker authentication logic in `ContainerManagerService` to explicitly utilize client configurations.
- Adjust logging levels and settings in `application-*.yml` files, including `okhttp3` logging.
- Handle missing or empty responses in `Web3RouterService` with default warnings and fallback values.

+++ src/main/java/io/hpp/noosphere/agent/service/blockchain/CoordinatorService.java
@@ -308,8 +308,8 @@ public class CoordinatorService {
         return Hash.sha3(buffer.array());
     }
 
-    public CompletableFuture<Boolean> hasRequestCommitments(BigInteger subscriptionId, BigInteger interval) {
-        byte[] requestId = getRequestId(subscriptionId, interval);
+    public CompletableFuture<Boolean> hasRequestCommitments(long subscriptionId, long interval) {
+        byte[] requestId = getRequestId(BigInteger.valueOf(subscriptionId), BigInteger.valueOf(interval));
         return web3DelegateeCoordinatorService
             .getRequestCommitment(requestId)
             .thenApply(commitment -> {
